# feat(rosbag2_to_non_annotated_t4): use cv2.remap instead of cv2.undistort