### Description

This PR refines the image undistortion process in the `_Rosbag2ToNonAnnotatedT4Converter` class:

- Removes the unused `_exec_undistort` flag.
- Computes undistortion maps (`undistort_map_x` and `undistort_map_y`) once when `undistort_image` is enabled, using `cv2.initUndistortRectifyMap()`. This avoids redundant calculations for every image frame.
- Replaces `cv2.undistort()` with `cv2.remap()` to utilize the precomputed undistortion maps, improving performance and maintaining consistent results.

### Notes

- The logic uses precomputed undistortion maps via `cv2.initUndistortRectifyMap`, which are then reused with `cv2.remap()` to optimize performance.

### Performance Improvement
Processed approximately 2,400 images (300 × 8 cameras).

#### ⏱️ Before
- Image undistortion time: **60.94 sec**
- **Total conversion time:** **168.10 sec**

#### ⚡ After
- Image undistortion time: **9.13 sec**
- **Total conversion time:** **105.82 sec**

#### 📉 Summary
- 🕒 Undistortion time reduced by ~85%
- 🧮 Total conversion time reduced by ~37%